pyproject.toml
[project]
name = "animal-shelter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "christavendel", email = "christavendel@ziggo.nl" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "matplotlib>=3.10.9",
    "pandas>=3.0.2",
    "pydantic-settings>=2.14.0",
    "uvicorn>=0.46.0",
]

[project.scripts]
animal-shelter = "animal_shelter:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "mypy>=1.20.2",
    "pandas-stubs>=3.0.0.260204",
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]

[tool.ruff.lint]
extend-select = [
  "UP", # pyupgrade
#  "D",  # pydocstyle
  "N",  # PEP8 names
  "I",  # isort
#  "TD"  # flake8 todos
]

[tool.ruff]
exclude = [
    ".git",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
     ".venv"
     ]
